requirements.txt
##NECESARIO PARA QUE FUNCIONE EL PROYECTO##
##COMANDO: pip install -r requirements.txt##
##pip install django-filter

django==3.2.7
pillow
django-admin-interface
django-crispy-forms
cx_Oracle
crispy_tailwind
django-filter